test(ipeps): move auto-bump integration test out of core marker
The 150s integration test landed in tests/test_chi_auto_bump.py, which
is mapped "core" in conftest.py. Pytest marker semantics are OR — even
with @pytest.mark.algorithm, the file-level "core" mark drags it into
CI's required -m core set, materially slowing every required check.

Move the integration test to its own file mapped "algorithm" so the
unit tests in test_chi_auto_bump.py stay in the core suite while the
end-to-end optimizer test runs only in the algorithm/full-suite path.
